fix: address copilot review on cache store
