+++ packages/core/lib/superdeck_core.dart
@@ -8,6 +8,7 @@ export 'src/models/block_model.dart';
 export 'src/models/deck_model.dart';
 export 'src/models/slide_model.dart';
 export 'src/asset_cache_store.dart';
+export 'src/asset_cache_store_io.dart';
 export 'src/deck_configuration.dart';
 export 'src/deck_service.dart';
 export 'src/deck_format_exception.dart';
